chore(auth): remove google signup button
refs #1305
