+++ src/components/Organization/Edit.tsx
@@ -7,9 +7,9 @@ import { FormProvider, SubmitHandler, useForm } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 import { useSaasAccount } from '~components/Account/SaasAccountProvider'
 import { ApiEndpoints } from '~components/Auth/api'
-import { HSeparator } from '~components/Auth/SignIn'
 import { useAuth } from '~components/Auth/useAuth'
 import { CreateOrgParams } from '~components/Organization/AccountTypes'
+import { HSeparator } from '~components/shared/Layout/Separators'
 import { DashboardBox } from '~shared/Dashboard/Contents'
 import FormSubmitMessage from '~shared/Layout/FormSubmitMessage'
 import { AvatarUploader } from '~shared/Layout/Uploader'
